Fixed errornous return values of characters that are above 0x7F.

+++ examples/performance/main.cpp
@@ -39,7 +39,7 @@ int32_t get_char(void * handle)
 {
     string_index_pair * p = static_cast<string_index_pair *>(handle);
     if (p->second < p->first.size()) {
-        return p->first[p->second++];
+        return static_cast<uint8_t>(p->first[p->second++]);
     }
     return -1;
 }
@@ -48,7 +48,7 @@ int32_t peek_char(void * handle)
 {
     string_index_pair * p = static_cast<string_index_pair *>(handle);
     if (p->second < p->first.size()) {
-        return p->first[p->second];
+        return static_cast<uint8_t>(p->first[p->second]);
     }
     return -1;
 }
